refactor: split reduce in new file
